refactor: create core module
