+++ eslint.config.js
@@ -9,6 +9,9 @@ export default defineConfig([
         extends: ["js/recommended"],
         languageOptions: {
             globals: globals.browser
+        },
+        rules: {
+            "no-unused-vars": "warn"
         }
     },
 ]);
