+++ src/core/canvas-app/CanvasAppEvent.js
@@ -1,0 +1,5 @@
+export class CanvasAppEvent {
+    constructor(app) {
+        this.app = app;
+    }
+}
