+++ src/core/canvas-app/CanvasAppPauseEvent.js
@@ -1,0 +1,8 @@
+import { CanvasAppEvent } from "./CanvasAppEvent.js";
+
+export class CanvasAppPauseEvent extends CanvasAppEvent {
+    constructor(app, isPaused) {
+        super(app);
+        this.isPaused = isPaused;
+    }
+}
