+++ src/core/index.js
@@ -1,0 +1,1 @@
+export * from "./canvas-app/index.js";
