--- src/events/AppEvents.js
@@ -1,12 +1,0 @@
-export class AppEvent {
-    constructor(app) {
-        this.app = app;
-    }
-}
-
-export class AppPauseEvent extends AppEvent {
-    constructor(app, isPaused) {
-        super(app);
-        this.isPaused = isPaused;
-    }
-}
