+++ src/index.js
@@ -1,4 +1,5 @@
-export * from "./events/index.js"
-export * from "./graphics/index.js"
-export * from "./math/index.js"
-export * from "./utils/index.js"
+export * from "./core/index.js";
+export * from "./events/index.js";
+export * from "./graphics/index.js";
+export * from "./math/index.js";
+export * from "./utils/index.js";
