+++ src/utils/index.js
@@ -1,3 +1,2 @@
-export * from "./CanvasApp.js";
 export * from "./Profiler.js";
 export * from "./Timer.js";
